Refactor Instances component out of Run component
